BJN-17: Moved context and hooks into their own files
