+++ Frontend/BlackJackNow/src/components/SocketMessage.tsx
@@ -1,12 +1,9 @@
 import { useState } from 'react';
-import { useSocket } from '../providers/Socketprovider';
-import { auth } from '../services/auth/firebaseAuthConfig';
+import { useSocket } from '../customHooks/useSocket';
 
 const SocketMessage = () => {
   const { socket } = useSocket();
   const [message, setMessage] = useState('');
-
-  console.log('Auth:', auth);
 
   if (!socket) {
     return <div>Loading...</div>;
