+++ Frontend/BlackJackNow/src/customHooks/useSocket.ts
@@ -1,0 +1,11 @@
+import { useContext } from 'react';
+import { SocketContext } from '../providers/SocketContext';
+
+export const useSocket = () => {
+  const context = useContext(SocketContext);
+  if (!context) {
+    throw new Error('useSocket must be used within a SocketProvider');
+  }
+  return context;
+};
+
