+++ Frontend/BlackJackNow/src/providers/SocketContext.ts
@@ -1,0 +1,13 @@
+import { createContext } from 'react';
+import { Socket } from 'socket.io-client';
+
+// Define the type for our context value
+export interface SocketContextType {
+  socket: Socket | null;
+}
+
+// Create the context with an initial value of null
+export const SocketContext = createContext<SocketContextType | undefined>(
+  undefined
+);
+
